Split static hostname stuff into its own file.
