+++ update-server/otupdate/common/name_management/static_hostname.py
@@ -1,0 +1,93 @@
+import asyncio
+import logging
+import os
+import urllib.parse
+
+
+LOG = logging.getLogger(__name__)
+
+
+async def set_up_static_hostname() -> str:
+    """Automatically configure the machine's static hostname.
+
+    Intended to be run once when the server starts.
+
+    This function:
+
+    1. Picks a good value for the static hostname.
+    2. Persists the new value for future boots.
+    3. Updates the "live" value for the current boot.
+    4. Restarts Avahi so that it advertises using the new live value.
+    5. Returns the new static hostname.
+    """
+    hostname = _choose_static_hostname()
+
+    with open("/etc/hostname", "w") as ehn:
+        ehn.write(f"{hostname}\n")
+
+    # First, we run hostnamed which will set the transient hostname
+    # and loaded static hostname from the value we just wrote to
+    # /etc/hostname
+    LOG.debug("Setting hostname")
+    proc = await asyncio.create_subprocess_exec(
+        "hostname",
+        hostname,
+        stdout=asyncio.subprocess.PIPE,
+        stderr=asyncio.subprocess.PIPE,
+    )
+    stdout, stderr = await proc.communicate()
+    ret = proc.returncode
+    if ret != 0:
+        LOG.error(
+            f"Error starting hostname: {ret} " f"stdout: {stdout!r} stderr: {stderr!r}"
+        )
+        raise RuntimeError("Couldn't run hostname")
+
+    # Then, with the hostname set, we can restart avahi
+    LOG.debug("Restarting avahi")
+    proc = await asyncio.create_subprocess_exec(
+        "systemctl",
+        "restart",
+        "avahi-daemon",
+        stdout=asyncio.subprocess.PIPE,
+        stderr=asyncio.subprocess.PIPE,
+    )
+    stdout, stderr = await proc.communicate()
+    ret = proc.returncode
+    if ret != 0:
+        LOG.error(
+            f"Error restarting avahi-daemon: {ret} "
+            f"stdout: {stdout!r} stderr: {stderr!r}"
+        )
+        raise RuntimeError("Error restarting avahi")
+    LOG.debug("Updated hostname and restarted avahi OK")
+
+    return hostname
+
+
+def _choose_static_hostname() -> str:
+    """Get a good value for the system's static hostname.
+
+    The static hostname is loaded from, in order of preference:
+
+    1. url-encoding the contents of /var/serial-number, if it is present and not empty.
+    2. the systemd-generated machine-id.
+    """
+    if os.path.exists("/var/serial"):
+        serial = open("/var/serial").read().strip()
+        if serial:
+            # TODO(mm, 2022-04-27): This uses the serial number even if it hasn't
+            # been configured and is still the default, like "opentrons."
+            LOG.info("Using serial for hostname")
+            hn = "".join([c for c in urllib.parse.quote(serial, safe="") if c != "%"])
+            if hn != serial:
+                LOG.warning(f"Reencoded serial to {hn}")
+            return hn
+
+        else:
+            LOG.info("Using machine-id for hostname: empty /var/serial")
+    else:
+        LOG.info("Using machine-id for hostname: no /var/serial")
+
+    with open("/etc/machine-id") as f:
+        return f.read().strip()[:6]
